Many minor changes

+++ src/include/generate/random.h
@@ -3,6 +3,8 @@
 
 #include <assert.h>
 
+#include <assert.h>
+
 class Random {
     static const long long multiplier = 0x5DEECE66DLL;
     static const long long addend = 0xBLL;
@@ -10,40 +12,28 @@ class Random {
     unsigned long long seed;
 
 public:
-    Random() { setSeed(140384); }
-
-    Random(long long s) { setSeed(s); }
-
-    void setSeed(long long s) { seed = (s ^ multiplier) & mask; }
-
-    int64_t next(int64_t bits) {
+    Random(){ setSeed(140384); }
+    Random(long long s){ setSeed(s); }
+    void setSeed(long long s){ seed = (s ^ multiplier) & mask; }
+    int next(int bits) {
         seed = (seed * multiplier + addend) & mask;
-        return (int) (seed >> (48 - bits));
+        return (int)(seed >> (48 - bits));
     }
-
-    int64_t nextInt(int n) {
+    int nextInt(int n) {
         assert(n > 0);
         // i.e., n is a power of 2
-        if ((n & -n) == n) return (int64_t) ((n * (long long) next(31)) >> 31);
-        int64_t bits, val;
-        do {
-            bits = next(31);
-            val = bits % n;
-        }
-        while (bits - val + (n - 1) < 0);
+        if ((n & -n) == n) return (int)((n * (long long)next(31)) >> 31);
+        int bits, val;
+        do { bits = next(31); val = bits % n; }
+        while (bits - val + (n-1) < 0);
         assert(val >= 0 && val < n);
         return val;
     }
-
-    int64_t nextInt() { return next(32); }
-
-    long long nextLong() { return ((long long) (next(32)) << 32) + next(32); }
-
-    bool nextBoolean() { return next(1) != 0; }
-
-    float nextFloat() { return next(24) / ((float) (1 << 24)); }
-
-    double nextDouble() { return (((long long) (next(26)) << 27) + next(27)) / (double) (1LL << 53); }
+    int nextInt(){ return next(32); }
+    long long nextLong(){ return ((long long)(next(32)) << 32) + next(32); }
+    bool nextBoolean(){ return next(1) != 0; }
+    float nextFloat(){ return next(24) / ((float)(1 << 24)); }
+    double nextDouble(){ return (((long long)(next(26)) << 27) + next(27)) / (double)(1LL << 53); }
 };
 
 #endif
